extracted multisig parsing to custom hook
