+++ src/providers/Fractal/governance/hooks/useSafeMultisigTxs.ts
@@ -1,13 +1,14 @@
-import { SafeMultisigTransactionWithTransfersResponse } from '@gnosis.pm/safe-service-client';
-import { format } from 'date-fns';
-import { BigNumber } from 'ethers';
 import { Dispatch, useEffect, useCallback } from 'react';
-import { formatWeiToValue } from '../../../../utils';
-import { DEFAULT_DATE_FORMAT } from '../../../../utils/numberFormats';
+import { useParseSafeTxs } from '../../../../hooks/utils/useParseSafeTxs';
 import { IGnosis } from '../../types';
-import { totalsReducer } from '../../utils';
 import { GovernanceAction, GovernanceActions } from '../actions';
-import { IGovernance, TxProposalState, GovernanceTypes, ActivityEventType } from './../types';
+import {
+  IGovernance,
+  TxProposalState,
+  GovernanceTypes,
+  ActivityEventType,
+  GovernanceActivity,
+} from './../types';
 interface IUseSafeMultisigTxs {
   governance: IGovernance;
   gnosis: IGnosis;
@@ -16,111 +17,23 @@ interface IUseSafeMultisigTxs {
 
 export const useSafeMultisigTxs = ({
   governance: { type },
-  gnosis: {
-    safeService,
-    transactions,
-    safe: { address },
-  },
+  gnosis: { transactions, safeService, safe },
   governanceDispatch,
 }: IUseSafeMultisigTxs) => {
+  const parsedActivities = useParseSafeTxs(transactions, safe);
+
   const getMultisigTx = useCallback(async () => {
-    if (!safeService || !address || !type) {
-      return;
-    }
-    if (!transactions.results.length) {
+    if (!safeService || !safe.address || !type) {
       return;
     }
 
-    const multisigTxs = transactions.results
-      .filter(tx => (tx as SafeMultisigTransactionWithTransfersResponse).safeTxHash)
-      .map((transaction, _, transactionArr) => {
-        const multiSigTransaction = transaction as SafeMultisigTransactionWithTransfersResponse;
+    if (!parsedActivities.length) {
+      return;
+    }
 
-        // @note for ethereum transactions event these are the execution date
-        const eventDate = format(new Date(multiSigTransaction.submissionDate), DEFAULT_DATE_FORMAT);
-
-        // returns mapping of Asset -> Asset Total Value by getting the total of each asset transfered
-        const transferAmountTotalsMap = transaction.transfers.reduce(totalsReducer, new Map());
-
-        // formats totals array into readable string with Symbol
-        const transferAmountTotals = Array.from(transferAmountTotalsMap.values()).map(token => {
-          const totalAmount = formatWeiToValue(token.bn, token.decimals);
-          const symbol = token.symbol;
-          return `${totalAmount} ${symbol}`;
-        });
-        const transferAddresses = transaction.transfers.map(transfer =>
-          transfer.to.toLowerCase() === address!.toLowerCase() ? transfer.from : transfer.to
-        );
-
-        const isEthSend =
-          !multiSigTransaction.data &&
-          !multiSigTransaction.isExecuted &&
-          !BigNumber.from(multiSigTransaction.value).isZero();
-
-        if (isEthSend) {
-          transferAmountTotals.push(`${formatWeiToValue(multiSigTransaction.value, 18)} ETHER`);
-          transferAddresses.push(multiSigTransaction.to);
-        }
-
-        const mappedTxHashes = transaction.transfers.map(transfer => transfer.transactionHash);
-
-        const txHashes = mappedTxHashes.length
-          ? mappedTxHashes
-          : [multiSigTransaction.transactionHash];
-
-        const eventSafeTxHash = multiSigTransaction.safeTxHash;
-
-        const eventNonce = multiSigTransaction.nonce;
-
-        const noncePair = transactionArr.find(tx => {
-          const multiSigTx = tx as SafeMultisigTransactionWithTransfersResponse;
-          return (
-            multiSigTx.nonce === eventNonce &&
-            multiSigTx.safeTxHash !== multiSigTransaction.safeTxHash
-          );
-        });
-
-        const isMultisigRejectionTx =
-          !multiSigTransaction.data &&
-          multiSigTransaction.to === multiSigTransaction.safe &&
-          noncePair &&
-          BigNumber.from(multiSigTransaction.value).isZero();
-
-        const isRejected = transactionArr.find(tx => {
-          const multiSigTx = tx as SafeMultisigTransactionWithTransfersResponse;
-          return (
-            multiSigTx.nonce === eventNonce &&
-            multiSigTx.safeTxHash !== multiSigTransaction.safeTxHash &&
-            multiSigTx.isExecuted
-          );
-        });
-
-        const isPending =
-          multiSigTransaction.confirmations?.length !== multiSigTransaction.confirmationsRequired;
-
-        const state = isRejected
-          ? TxProposalState.Rejected
-          : isPending
-          ? TxProposalState.Pending
-          : !multiSigTransaction.isExecuted
-          ? TxProposalState.Active
-          : multiSigTransaction.isSuccessful && multiSigTransaction.isExecuted
-          ? TxProposalState.Executed
-          : TxProposalState.Pending;
-
-        return {
-          state,
-          eventType: ActivityEventType.Governance,
-          eventDate, // update this
-          proposalNumber: eventSafeTxHash,
-          targets: [transaction.to],
-          txHashes,
-          multisigRejectedProposalNumber:
-            isMultisigRejectionTx && !!noncePair
-              ? (noncePair as SafeMultisigTransactionWithTransfersResponse).safeTxHash
-              : undefined,
-        };
-      });
+    const multisigTxs = (parsedActivities as GovernanceActivity[]).filter(
+      tx => tx.eventType === ActivityEventType.Governance
+    );
 
     const passedProposals = multisigTxs.reduce(
       (prev, proposal) => (proposal.state === TxProposalState.Executed ? prev + 1 : prev),
@@ -143,7 +56,7 @@ export const useSafeMultisigTxs = ({
         pending: pendingProposals,
       },
     });
-  }, [safeService, address, governanceDispatch, type, transactions]);
+  }, [safeService, safe.address, governanceDispatch, type, parsedActivities]);
 
   useEffect(() => {
     if (type === GovernanceTypes.GNOSIS_SAFE) {
